Gyanodaya

+++ src/Screens/cmsScreen/cms-components/cms-gallery/galleryApii.js
@@ -1,4 +1,5 @@
 import axios from "axios";
+const getAuthToken = () => localStorage.getItem("authToken");
 
 const BASE_URL = import.meta.env.VITE_API_URL;
 
@@ -13,27 +14,30 @@ export const getGalleryById = async (id) => {
 };
 
 export const updateGalleryById = async (id, data) => {
-    const token = localStorage.getItem('authToken')
+    const token = getAuthToken();
+    if (!token) throw new Error("Token is missing");
     const headers = {
-        'Authorization': `Bearer ${token}`
+      Authorization: `Bearer ${token}`,
     };
     const response = await axios.patch(`${BASE_URL}/gallery/${id}`, data, {headers});
     return response.data;
 };
 
 export const addGallery = async (data) => {
-    const token = localStorage.getItem('authToken')
+    const token = getAuthToken();
+    if (!token) throw new Error("Token is missing");
     const headers = {
-        'Authorization': `Bearer ${token}`
+      Authorization: `Bearer ${token}`,
     };
     const response = await axios.post(`${BASE_URL}/gallery`, data, {headers});
     return response.data;
 };
 
 export const deleteGallery = async (id) => {
-    const token = localStorage.getItem('authToken')
+    const token = getAuthToken();
+    if (!token) throw new Error("Token is missing");
     const headers = {
-        'Authorization': `Bearer ${token}`
+      Authorization: `Bearer ${token}`,
     };
     const response = await axios.delete(`${BASE_URL}/gallery/${id}`, {headers});
     return response.data;
